Beginnings of an App
Our app starts out in the file `config.ru`, which our `rackup` command
looks for by default when it's run.

We define an Application class with a `call` method on it. Each time a
web request is received from a browser, Rack will call `call`.

Our `call` method returns a specially formatted array that Rack is
expecting.