Take a name from the query string
Plain output is kind of boring, so let's respond to our users a bit.

With this commit, if a user passes us a query string including a name,
we'll use that to make a friendlier message. A query string is the
portion of a URL after the ?, so if a user visited
http://localhost:9292?name=Jason, then the query string is "name=Jason".

We use a class that Rack provides called Rack::Request to parse the
incoming env data. The query string is available to us as a Ruby hash
via the `params` method.