HTML... with templating!
The web runs on HTML, so let's update our app to output some real HTML
instead of just plain text.

page.html is what we refer to as a "template" file. It has a placeholder
in it that we will plug our data into, in this case `<%=name%> which
gets fed the `name` variable we've read in the call method.

We do the templating with ERB, a library for just this purpose that
comes with Ruby itself. If the ERB invocation looks a little peculiar,
don't stress! The important bit about it is that the `binding` we pass
to it let's the template get access to all our local variables. Cool!