An input form, but where to?
This commit adds a simple input form on our start page. Users can enter
text, and when they click the button it's submitted to the server. (You
can tell because the browser refreshes)

Notice that the value's cleared, and we don't yet do anything with the
data submitted. We'll fix that next!