Update our greeting with what the users posts
You may have noticed our form included a bit saying `method=post`. This
tells the browser to send our data back in as a POST rather than a GET
(like our initial render of the page).

We use the `Rack::Request` methods `get?` and `post?` to see whether
we're doing our initial render, or responding to a POST. If we're a
POST, then we factor in the user's input to us.