Extract content templating
To prepare ourselves for the future a bit, we want to extract a few
methods from our `call`. The first of these is generating the ERB
content from a file and our list of values to template in.

When our ERB call moves outside our current method, though, we lose the
local values to pass with the binding. How do we get past this hurdle?

Luckily, Ruby's Binding class will let us attach additional locals to a
binding by the `local_variable_set` method. We update our method to take
a parameter hash, and put each of those values onto the binding before
running it through ERB.