Add a simple middleware
You might have thought that Rack's method `call` on the `Application`
class seemed a little generically named... Why not `handle_request` or
`render` or something like that? Well, the naming's on purpose, and one
reason is that you can define other objects which respond to `call` that
we refer to as middleware.

Middleware classes defined in a chain by calls to `use` in your
config.ru. They see your web request before your application class, and
then see your response on the way out after.

To start with, we simply write a little output to the console to prove
to ourselves that our middleware is getting called. You should see
something like this when you browse a page running this version of the
code:

  Hi there!
  127.0.0.1 - - [19/Aug/2015:22:11:26 -0700] "GET /?name=Jason HTTP/1.1" 200 - 0.0014

Nothing too exciting, but don't worry, we'll do something cooler with it
in the next commit!