Add a debugging middleware
The interesting thing about middleware is that it gets to see our
request and response both before and after the main application object
has decided how to respond.

We use that power to add some debugging information. If `debug=1`
is on the query string, we modify the outgoing response to include our
request parameters. That's right, we modified the response without the
`Application` class having to know anything about it!