Gone 404
As alluded to before, most websites don't serve you a page for just any
path that you happen to request on it. They're probably onto something,
so let's shore up our little app to only respond to our two URL's:
/farewell and /.

You might remember in the early going when we returned 200 as the first
element in the array from our `call` method? Well now we return with 404
if we're not a recognized path. That's HTTP for "nothing here, move
along now."