Routing!
Here we finally see the first bit of payoff from our strange refactors.
All those Proc's that we were handling in our carefully crafted if
statements are now bundled up into a neat Hash. Our HTTP method and path
become the keys, the Proc the value we get back.

Some noise comes along with this, though. We have to pass the
`Application` instance into the Proc's for them to be able to call
methods on them. Plus the Hash syntax, while nicer than our nested if's,
still isn't that clean to look at.

Let's see if we can do better.