Cleaner declaration for routes
So one of our complaints was that the Hash didn't make for a very
intuitive place to wire up our various Proc's. This commit fixes that by
introducing a small DSL in it's place. The underlying data is the same,
but now we can use a far more satisfying `get` and `post` call instead
of those odd key Arrays.