Farewell passing the app
The other big complaint we had with our rudamentary router was the fact
that we had to pass our `Application` instance in.

Ruby metaprogramming to the rescue!

Now our `get` and `post` blocks can assume the context of running with
an `Application` instance, without needing an extra parameter. How do we
accomplish this? Through the magic of `instance_exec`, which allows you
to run an arbitrary block of Ruby code as if it were a method defined
inside of an object.